Refactor comment utilities into focused modules
